Fix #39. Refactoring.
