+++ internal/domain/dao/balances.go
@@ -1,5 +1,5 @@
 /*
- * This file was last modified at 2024-04-19 15:58 by Victor N. Skurikhin.
+ * This file was last modified at 2024-04-20 17:50 by Victor N. Skurikhin.
  * balances.go
  * $Id$
  */
@@ -8,8 +8,8 @@ package dao
 
 import (
 	"github.com/vskurikhin/gophermart/internal/domain/entity"
+	"github.com/vskurikhin/gophermart/internal/domain/view"
 	"github.com/vskurikhin/gophermart/internal/storage"
-	"math/big"
 )
 
 type balances struct {
@@ -20,18 +20,10 @@ func Balances(storage storage.Storage) *balances {
 	return &balances{storage: storage}
 }
 
-func (b *balances) GetAllBalances() ([]*entity.Balance, error) {
-	return entity.FuncGetAllBalances()(b.storage)
-}
-
 func (b *balances) GetBalance(login string) (*entity.Balance, error) {
 	return entity.FuncGetBalance()(b.storage, login)
 }
 
-func (b *balances) GetBalanceWithdraw(login string) (*entity.Balance, *big.Float, error) {
-	return entity.FuncGetBalanceWithdraw()(b.storage, login)
-}
-
-func (b *balances) Save(balance *entity.Balance) (*entity.Balance, error) {
-	return balance.Save(b.storage)
+func (b *balances) GetBalanceWithdraw(login string) (*view.Balance, error) {
+	return view.FuncGetBalanceWithdraw()(b.storage, login)
 }
