+++ internal/domain/dao/orders.go
@@ -1,5 +1,5 @@
 /*
- * This file was last modified at 2024-04-20 17:09 by Victor N. Skurikhin.
+ * This file was last modified at 2024-04-20 17:46 by Victor N. Skurikhin.
  * orders.go
  * $Id$
  */
@@ -20,16 +20,8 @@ func Orders(storage storage.Storage) *orders {
 	return &orders{storage: storage}
 }
 
-func (o *orders) GetAllOrders() ([]*entity.Order, error) {
-	return entity.FuncGetAllOrders()(o.storage)
-}
-
 func (o *orders) GetAllOrdersForLogin(login string) ([]*view.Order, error) {
 	return view.FuncGetAllOrdersForLogin()(o.storage, login)
-}
-
-func (o *orders) GetOrder(login, number string) (*entity.Order, error) {
-	return entity.FuncGetOrder()(o.storage, login, number)
 }
 
 func (o *orders) GetOrderByNumber(number string) (*entity.Order, error) {
@@ -38,8 +30,4 @@ func (o *orders) GetOrderByNumber(number string) (*entity.Order, error) {
 
 func (o *orders) Insert(order *entity.Order) (*entity.Order, error) {
 	return order.Insert(o.storage)
-}
-
-func (o *orders) Save(order *entity.Order) (*entity.Order, error) {
-	return order.Save(o.storage)
 }
