+++ internal/domain/dao/users.go
@@ -1,5 +1,5 @@
 /*
- * This file was last modified at 2024-04-15 15:07 by Victor N. Skurikhin.
+ * This file was last modified at 2024-04-20 17:46 by Victor N. Skurikhin.
  * users.go
  * $Id$
  */
@@ -19,18 +19,6 @@ func Users(storage storage.Storage) *users {
 	return &users{storage: storage}
 }
 
-func (u *users) GetAllUsers() ([]*entity.User, error) {
-	return entity.FuncGetAllUsers()(u.storage)
-}
-
 func (u *users) GetUser(login string) (*entity.User, error) {
 	return entity.FuncGetUser()(u.storage, login)
-}
-
-func (u *users) Insert(user *entity.User) (*entity.User, error) {
-	return user.Insert(u.storage)
-}
-
-func (u *users) Save(user *entity.User) (*entity.User, error) {
-	return user.Save(u.storage)
 }
