+++ internal/domain/entity/order.go
@@ -1,5 +1,5 @@
 /*
- * This file was last modified at 2024-04-20 17:09 by Victor N. Skurikhin.
+ * This file was last modified at 2024-04-20 17:46 by Victor N. Skurikhin.
  * order.go
  * $Id$
  */
@@ -28,28 +28,8 @@ func NewOrder(login string, number string) *Order {
 	return &Order{login: login, number: number}
 }
 
-func NewOrderStatus(login string, number string, statusID int) *Order {
-	return &Order{login: login, number: number, statusID: statusID}
-}
-
 func (o *Order) Login() string {
 	return o.login
-}
-
-func (o *Order) Number() string {
-	return o.number
-}
-
-func (o *Order) SetNumber(number string) {
-	o.number = number
-}
-
-func (o *Order) StatusID() int {
-	return o.statusID
-}
-
-func (o *Order) SetStatusID(statusID int) {
-	o.statusID = statusID
 }
 
 func (o *Order) Accrual() *big.Float {
@@ -58,18 +38,6 @@ func (o *Order) Accrual() *big.Float {
 
 func (o *Order) SetAccrual(accrual *big.Float) {
 	o.accrual = accrual
-}
-
-func (o *Order) UploadedAt() *time.Time {
-	return o.uploadedAt
-}
-
-func (o *Order) CreatedAt() time.Time {
-	return o.createdAt
-}
-
-func (o *Order) UpdateAt() *time.Time {
-	return o.updateAt
 }
 
 func (o *Order) Insert(s storage.Storage) (*Order, error) {
@@ -94,38 +62,6 @@ func (o *Order) Insert(s storage.Storage) (*Order, error) {
 			o.login, o.number,
 		)
 	}
-	if err != nil {
-		return nil, err
-	}
-	return extractOrder(row)
-}
-
-func (o *Order) Save(s storage.Storage) (*Order, error) {
-
-	var accrualNullFloat64 sql.NullFloat64
-	var uploadedAtNullTime sql.NullTime
-
-	if o.accrual != nil {
-		accrual, _ := (*o).accrual.Float64()
-		accrualNullFloat64.Float64 = accrual
-		accrualNullFloat64.Valid = true
-	}
-	if o.uploadedAt != nil {
-		uploadedAtNullTime.Time = *o.uploadedAt
-		uploadedAtNullTime.Valid = true
-	}
-	row, err := s.Save(
-		`INSERT INTO "orders"
-				    (login, number, status_id, accrual, uploaded_at, created_at)
-             VALUES ($1, $2, $3, $4, $5, now())
-             ON CONFLICT (login, number)
-             DO UPDATE SET
-               status_id = $3,
-			   accrual = $4,
-               uploaded_at = $5
-             RETURNING *`,
-		o.login, o.number, o.statusID, accrualNullFloat64, uploadedAtNullTime,
-	)
 	if err != nil {
 		return nil, err
 	}
@@ -157,33 +93,12 @@ func (o *Order) AppendSetStatusTo(a storage.TxArgs, status string) storage.TxArg
 	return append(a, t)
 }
 
-func FuncGetAllOrders() func(storage.Storage) ([]*Order, error) {
-	return func(s storage.Storage) ([]*Order, error) {
-		result := make([]*Order, 0)
-		return result, nil
-	}
-}
-
 func FuncGetOrderByNumber() func(storage.Storage, string) (*Order, error) {
 	return func(s storage.Storage, number string) (*Order, error) {
 
 		row, err := s.GetByNumber(
 			`SELECT * FROM "orders" WHERE number = $1`,
 			number,
-		)
-		if err != nil {
-			return nil, err
-		}
-		return extractOrder(row)
-	}
-}
-
-func FuncGetOrder() func(storage.Storage, string, string) (*Order, error) {
-	return func(s storage.Storage, login, number string) (*Order, error) {
-
-		row, err := s.GetByLoginNumber(
-			`SELECT * FROM "orders" WHERE login = $1 AND number = $2`,
-			login, number,
 		)
 		if err != nil {
 			return nil, err
