+++ internal/handlers/accounts/balance_withdraw.go
@@ -1,5 +1,5 @@
 /*
- * This file was last modified at 2024-04-20 00:34 by Victor N. Skurikhin.
+ * This file was last modified at 2024-04-20 18:06 by Victor N. Skurikhin.
  * balance_withdraw.go
  * $Id$
  */
@@ -28,6 +28,7 @@ func newBalanceWithdraw() *balanceWithdraw {
 	return &balanceWithdraw{log: logger.Get()}
 }
 
+//goland:noinspection GoUnhandledErrorResult
 func (r *balanceWithdraw) Handle(response http.ResponseWriter, request *http.Request) {
 
 	ctx := request.Context()
@@ -37,7 +38,6 @@ func (r *balanceWithdraw) Handle(response http.ResponseWriter, request *http.Req
 
 		r.log.Debug(balanceWithdrawMsg, utils.LogCtxRecoverFields(ctx, err)...)
 		render.Status(request, http.StatusBadRequest)
-		//goland:noinspection GoUnhandledErrorResult
 		render.Render(response, request, model.Error(handlers.ErrBadRequest))
 
 		return
@@ -48,26 +48,21 @@ func (r *balanceWithdraw) Handle(response http.ResponseWriter, request *http.Req
 
 		r.log.Debug(balanceWithdrawMsg, utils.LogCtxRecoverFields(ctx, err)...)
 		render.Status(request, http.StatusBadRequest)
-		//goland:noinspection GoUnhandledErrorResult
 		render.Render(response, request, model.Error(handlers.ErrBadRequest))
 
 		return
 	}
 	result := newService(ctx, storage.NewPgsStorage()).Withdraw(*login, withdraw)
+	render.Status(request, result.Status())
 
 	switch value := result.(type) {
 	case *handlers.ResultError:
-
-		render.Status(request, value.Status())
-		//goland:noinspection GoUnhandledErrorResult
 		render.Render(response, request, model.Error(value.Error()))
-
 	case *handlers.ResultAny:
 
 		if withdraw, ok := value.Result().(*model.Withdraw); ok {
 			response.Header().Set("Content-Type", "application/json")
 			if err := withdraw.MarshalToWriter(response); err == nil {
-				render.Status(request, value.Status())
 				return
 			} else {
 				r.log.Debug(balanceWithdrawMsg, utils.LogCtxReasonErrFields(ctx, err.Error(), handlers.ErrInternalError)...)
