+++ internal/handlers/orders/orders.go
@@ -1,5 +1,5 @@
 /*
- * This file was last modified at 2024-04-20 17:09 by Victor N. Skurikhin.
+ * This file was last modified at 2024-04-20 18:04 by Victor N. Skurikhin.
  * orders.go
  * $Id$
  */
@@ -29,6 +29,7 @@ func newOrders() *orders {
 	return &orders{log: logger.Get()}
 }
 
+//goland:noinspection GoUnhandledErrorResult
 func (o *orders) Handle(response http.ResponseWriter, request *http.Request) {
 
 	ctx := request.Context()
@@ -38,19 +39,16 @@ func (o *orders) Handle(response http.ResponseWriter, request *http.Request) {
 
 		o.log.Debug(ordersMsg, utils.LogCtxRecoverFields(ctx, err)...)
 		render.Status(request, http.StatusBadRequest)
-		//goland:noinspection GoUnhandledErrorResult
 		render.Render(response, request, model.Error(handlers.ErrBadRequest))
 
 		return
 	}
 	result := newService(ctx, storage.NewPgsStorage(), accrual.GetWorkers()).
 		Orders(*login)
+	render.Status(request, result.Status())
 
 	switch value := result.(type) {
 	case *handlers.ResultError:
-
-		render.Status(request, value.Status())
-		//goland:noinspection GoUnhandledErrorResult
 		render.Render(response, request, model.Error(value.Error()))
 
 	case *handlers.ResultAny:
@@ -58,7 +56,6 @@ func (o *orders) Handle(response http.ResponseWriter, request *http.Request) {
 		if orders, ok := value.Result().(model.Orders); ok {
 			response.Header().Set("Content-Type", "application/json")
 			if err := orders.MarshalToWriter(response); err == nil {
-				render.Status(request, value.Status())
 				return
 			} else {
 				o.log.Debug(ordersMsg, utils.LogCtxReasonErrFields(ctx, err.Error(), handlers.ErrInternalError)...)
